Listen on all interfaces by defaults in Docker
Normally, jiralerts appears to be configured to listen on the loopback
interface by default. This seems like a sane default, as it allows (by
proxy) pushing arbitrary data into a Jira installation, without any
additional authentication except that supplied at runtime.

However, in the case of Docker (and in particular Kubernetes) the
interface that will be queried for a response is not the loopback, but
rather another bridge interface.

This commit modifies the start command of the container to listen on all
interfaces by default.