Correct the path to the jiralerts application
Initially, when the Dockerfile was created, the jiralerts application
was thought to be called "jiraalerts" -- or, "jira + alerts". However,
this is incorrect.

This commit fixes the paths in the container to reflect this change.

Additionally, it splits out adding main.py from adding the requirements.
This has no impact in the container construction (where `docker build
--squash is used) but allows a faster development / deployment cycle
when iterating on the script), thanks to Dockers caching of the
filesystem layers.