Make compatible with the newer syntax used by the alert manager
Recently, changes were made to the alert manager output such that

1. The version is now "4"
2. The "group key" is now a string, of arbitrary characters

Previously, the group key was number, and could be supplied to
Jira in the form of JQL without any escaping. However, now that it is a
string, it includes characters such as the "{}" character or spaces that
have special meaning within JQL.

Group key is still a good candidate for matching tickets. Thus, it has
been base64 encoded and that is used as the matching key. This does not
lose any semantic information (at least, it can be reversed) but is
compatible with JQL now and going forward.

BREAKING CHANGE:

This will break matching for any existing tickets that are using
the numeric ticket reference to keep track.