Make the initial task configurable
Currently, when an alert arrives a new task of the type "Task" is
created. However, this type does not always exist in projects; in the
authors example, the types:

  - "Bug"
  - "Story"
  - "Epic"

exist, but the type "Task" does not. This commit makes the initial task
configurable, but leaves the default value of "Task"

Design Notes:

configparser:

  The configuration format specified by the "configparser" python stdlib
  is used for configuration. This was used as the author has very
  limited experience with python, but presumes the configparser is the
  most common way of expressing configuration.

  Future work might consider merging command line variables into this
  ini format.