Settle on a different method of providing defaults
Previously, work was implemented such that a configuration file could be
used to dictate what sort of task should be created with a new alert.
However, the author wishes the application to work entirely without the
configuration file.

This commit modifies the configuration such that the file is no longer
required to supply default values, and instead is optional. The
configuration defaults are supplied inline.